preprocessor/requirements.txt
fastapi[all]
pydantic
uvicorn[standard]
setuptools
sqlalchemy==2.0.31
sqlacodegen==3.0.0rc5
python-dotenv
tiktoken
langchain
langchain-openai
langchain-text-splitters
requests
bs4
sentence_transformers
unstructured
openai
pgvector
psycopg2
unstructured[all-docs]
pillow
pydantic
lxml
pillow
matplotlib
pdf2image
markdownify
psutil
